# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Rare District — Nigerian fashion marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [Orders]
      summary: List current user's orders
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"

    post:
      operationId: createOrder
      tags: [Orders]
      summary: Create an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [Orders]
      summary: Get order details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [Orders]
      summary: Update order status (vendor or admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # ── PAYMENTS ──────────────────────────────────────────────────────────────
  /payments/paystack/initiate:
    post:
      operationId: initiatePaystackPayment
      tags: [Payments]
      summary: Initiate a Paystack payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInit"
      responses:
        "200":
          description: Payment URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentInitResponse"
  /payments/flutterwave/initiate:
    post:
      operationId: initiateFlutterwavePayment
      tags: [Payments]
      summary: Initiate a Flutterwave payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInit"
      responses:
        "200":
          description: Payment URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentInitResponse"
  /payments/paystack/verify:
    post:
      operationId: verifyPaystackPayment
      tags: [Payments]
      summary: Verify a Paystack payment by reference
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentVerify"
      responses:
        "200":
          description: Payment result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentResult"
  /payments/flutterwave/verify:
    post:
      operationId: verifyFlutterwavePayment
      tags: [Payments]
      summary: Verify a Flutterwave payment by reference
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentVerify"
      responses:
        "200":
          description: Payment result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentResult"

  # ── COUPONS ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    User:
      type: object
      required: [id, email, role, referralCode, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        role:
          type: string
          enum: [shopper, vendor, admin]
        referralCode:
          type: string
        referredBy:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    Vendor:
      type: object
      required: [id, userId, brandName, status, payoutBalance, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        brandName:
          type: string
        description:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        bankName:
          type: ["string", "null"]
        accountNumber:
          type: ["string", "null"]
        accountName:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, rejected]
        commissionRateOverride:
          type: ["number", "null"]
        payoutBalance:
          type: number
        adminNote:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
...
    Product:
      type: object
      required: [id, vendorId, name, price, currency, stock, isActive, createdAt]
      properties:
        id:
          type: integer
        vendorId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        currency:
          type: string
          default: NGN
        category:
          type: ["string", "null"]
        sizes:
          type: array
          items:
            type: string
        images:
          type: array
          items:
            type: string
        stock:
          type: integer
        isActive:
          type: boolean
        isFeatured:
          type: boolean
        wardrobeCount:
          type: integer
        averageRating:
          type: ["number", "null"]
        reviewCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        vendor:
          $ref: "#/components/schemas/Vendor"
...
    Order:
      type: object
      required: [id, userId, status, totalAmount, currency, shippingAddress, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        status:
          type: string
          enum: [pending, paid, processing, shipped, delivered, cancelled]
        totalAmount:
          type: number
        discountAmount:
          type: number
        currency:
          type: string
          default: NGN
        shippingAddress:
          type: string
        shippingCity:
          type: string
        shippingState:
          type: string
        shippingPhone:
          type: string
        couponCode:
          type: ["string", "null"]
        paymentProcessor:
          type: ["string", "null"]
          enum: [paystack, flutterwave, null]
        paymentReference:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
    OrderItem:
      type: object
      required: [id, orderId, productId, vendorId, quantity, unitPrice, commissionAmount, vendorAmount]
      properties:
        id:
          type: integer
        orderId:
          type: integer
        productId:
          type: integer
        vendorId:
          type: integer
        quantity:
          type: integer
        selectedSize:
          type: ["string", "null"]
        unitPrice:
          type: number
        commissionRate:
          type: number
        commissionAmount:
          type: number
        vendorAmount:
          type: number
        product:
          $ref: "#/components/schemas/Product"
    OrderInput:
      type: object
      required: [items, shippingAddress, shippingCity, shippingState, shippingPhone]
      properties:
        items:
          type: array
          minItems: 1
          items:
            type: object
            required: [productId, quantity]
            properties:
              productId:
                type: integer
              quantity:
                type: integer
                minimum: 1
              selectedSize:
                type: string
        shippingAddress:
          type: string
        shippingCity:
          type: string
        shippingState:
          type: string
        shippingPhone:
          type: string
        couponCode:
          type: string
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [processing, shipped, delivered, cancelled]

    # ── Payments ──────────────────────────────────────────────────────────────
    PaymentInit:
      type: object
      required: [orderId, email]
      properties:
        orderId:
          type: integer
        email:
          type: string

        callbackUrl:
          type: string
    PaymentInitResponse:
      type: object
      required: [paymentUrl, reference]
      properties:
        paymentUrl:
          type: string
        reference:
          type: string
    PaymentVerify:
      type: object
      required: [reference]
      properties:
        reference:
          type: string
    PaymentResult:
      type: object
      required: [success, status, orderId]
      properties:
        success:
          type: boolean
        status:
          type: string
        orderId:
          type: integer
        amount:
          type: number
        reference:
          type: string

    # ── Coupons ───────────────────────────────────────────────────────────────
...