# lib/api-zod/src/generated/types/paymentInit.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Rare District — Nigerian fashion marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface PaymentInit {
  orderId: number;
  email: string;
  callbackUrl?: string;
}